Add codebuild role policy with s3:GetObject permissions
We think this is needed so the pipeline can build the API Gateway using
terraform. The S3 get object is needed to access the tfstate file in the
integration bucket.

Note, that we still have a permissions issue. We think the next step is
to add a policy to the ci role.

Co-authored-by: Rachel Howell <98879818+rachowell@users.noreply.github.com>