Create cognito user pool and api_gateway authorizer
Ticket: EC-60

Note that CloudFront still needs to be configured to pass through the
Authorization header. We are currently using the User Pool's Hosted UI
to get an authorization code with which we can then get an access token.
Instead, we want to change the auth flow so we can hit an endpoint with
 username/password and receive a token back directly.

Co-Authored-By: Rachel Howell <98879818+rachowell@users.noreply.github.com>
Co-Authored-By: Russ Penska <russ-madetech@users.noreply.github.com>