Add "api." prefix to domain for the api
Trying to apply Terraform changes without the "api." prefix resulted in an
error from AWS that the domain isn't covered by the SSL certificate.

The certificate is for *.building-energy-calculator-integration.communities.gov.uk, so seems like it needs a prefix (and this is what we planned for the full URL for the service).