Comment out Cloudfront SSL cert attributes until custom cert
Terraform wants to make changes all the time as the
minimum_protocol_version and ssl_support_method keys
only apply when using a custom certificate.

Commenting these out for now so there are no TF changes on main branch,
while we wait for DNS records to be made for provisioning the custom certificate.