Rename frontend pipeline to disambiguate from other frontend pipelines
In the main register CI account we have both "epbr-frontend-pipeline" and "epbr-data-frontend-pipeline".  When collecting CCTray status updates from both ECaaS and register CI accounts the name "front-end-pipeline" is ambiguous.  By including "ecaas" in the pipeline name it will be easier to know where to look when there is a build error.