REFACTOR: Move Cognito into its own module and use SSL module for frontend
Cognito is used by both api_gateway and front_end modules so shouldn't be inside api_gateway.  This refactors it out into its own module with outputs that can be referenced in params for both api_gateway and front_end modules.

The front_end module has its own certificate internally but it's better to re-use the ssl module that already exists for consistency.

The S3 bucket used by the front end in integration doesn't include the environment name (or anything variable) so won't be able to be built in other environments due to a name clash. This add the "environment" variable into the name so we get a unique bucket name per environment.

For now a few changes are prevented from being applied in integration with `var.environment == "ecaas-intg" ?` conditionals.  This is because we don't want to risk breaking integration right now, but once we have further environments built we'll remove these conditionals and re-build integration according to the new naming conventions.