Remove front_end_lambda.code_sha256 as a trigger for API Gateway redeployment
We shouldn't need to redeploy the API Gateway every time the frontend Lambda code is updated.

Including this in the redeployment triggers causes Terraform to monitor the value of this code_sha256 and gives a warning that "Note: Objects have changed outside of Terraform" if the frontend code has changed. We don't want Terraform to monitor the exact version of the frontend (or API) code that's deployed, as these are managed by CICD, not Terraform.