Don't require subject_alternative_names for SSL certificates
Our SSL certs for ECaaS are one domain (plus a wildcard as a subject alternative name), so no need to pass these in as a separate list.

Including this as a separate parameter was a throwback to the register services where we have multiple legacy domains listed on the same SSL cert in some instances.